Initial refactoring and disabled security check for https

+++ main.go
@@ -1,28 +1,40 @@
 package main
 
 import (
-	"os"
-	"log"
 	"bufio"
-	"fmt"
-	"sync"
-	_ "github.com/mattn/go-sqlite3"
+	"crypto/tls"
 	"database/sql"
-	"runtime"
 	"flag"
+	"fmt"
+	"log"
 	"net/http"
+	"os"
+	"runtime"
+	"sync"
+	"time"
+
+	_ "github.com/mattn/go-sqlite3"
 )
 
 func requestWorker(domains chan string, wg *sync.WaitGroup, db *sql.DB) {
 	defer wg.Done()
+	//defer db.Close()
 
-	httpClient := &http.Client{}
+	//use this to disable security check validation
+	tr := &http.Transport{
+		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+	}
+
+	httpClient := &http.Client{Transport: tr}
+
+	//httpClient := &http.Client{}
 
 	for domain := range domains {
-		response, err := httpClient.Get(domain)
-		var statusCode int;
+		response, err := httpClient.Head(domain)
 
-		if (err == nil) {
+		var statusCode int
+
+		if err == nil {
 			statusCode = response.StatusCode
 			log.Println(fmt.Sprintf("[%s] %d", domain, response.StatusCode))
 		} else {
@@ -30,9 +42,12 @@ func requestWorker(domains chan string, wg *sync.WaitGroup, db *sql.DB) {
 			log.Println(fmt.Sprintf("[%s] HTTP Error %s", domain, err.Error()))
 		}
 
-		go func() {
-			db.Exec(fmt.Sprintf("INSERT INTO sites(domain, code) values(\"%s\", %d)", domain, statusCode))
-		}()
+		//log.Println(fmt.Sprintf("INSERT INTO sites(id, domain, code) values(\"%s\", %d)", domain, statusCode))
+		_, err = db.Exec(fmt.Sprintf("INSERT INTO sites(domain, code, time) values(\"%s\", %d, \"%s\")", domain, statusCode, time.Now()))
+		if err != nil {
+			log.Fatalln("could not insert row:", err)
+		}
+
 	}
 }
 
@@ -41,37 +56,52 @@ func initDataBase() *sql.DB {
 	if err != nil {
 		log.Fatal(err)
 	}
-	defer db.Close()
 
-	sqlStmt := `
-	DELETE FROM sites;
-	DROP TABLE sites;
-	`
-	db.Exec(sqlStmt)
+	rows, err := db.Query("SELECT name FROM sqlite_master WHERE type='table' AND name='sites'")
 	if err != nil {
-		log.Panic("%q: %s\n", err, sqlStmt)
+		log.Panic("%q: \n", err)
 		os.Exit(1)
 	}
+	defer rows.Close()
+	var name string
 
-	sqlStmt = `
-	CREATE TABLE sites (domain TEXT NOT NULL PRIMARY KEY, code INTEGER);
-	`
-	_, err = db.Exec(sqlStmt)
-	if err != nil {
-		log.Panic("%q: %s\n", err, sqlStmt)
-		os.Exit(1)
+	for rows.Next() {
+		rows.Scan(&name)
+		//log.Println(name)
 	}
 
+	if name != "sites" {
+
+		sqlStmt := `
+			DELETE FROM sites;
+			DROP TABLE sites;
+			`
+		db.Exec(sqlStmt)
+		if err != nil {
+			log.Panic("%q: %s\n", err, sqlStmt)
+			os.Exit(1)
+		}
+
+		sqlStmt = `
+			CREATE TABLE sites (site_id INTEGER PRIMARY KEY AUTOINCREMENT, domain TEXT NOT NULL, code INTEGER, time TEXT NOT NULL);
+			`
+		_, err = db.Exec(sqlStmt)
+		if err != nil {
+			log.Panic("%q: %s\n", err, sqlStmt)
+			os.Exit(1)
+		}
+	}
 
 	return db
 }
 func main() {
+
 	runtime.GOMAXPROCS(runtime.NumCPU())
 
-	var threadsCount int;
+	var threadsCount int
 	flag.IntVar(&threadsCount, "threads", 15, "Count of used threads (goroutines)")
 
-	file, err := os.Open("ru_domains_200_ok")
+	file, err := os.Open("domains_tc")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -82,7 +112,7 @@ func main() {
 	taskChannel := make(chan string, 10000)
 	go func() {
 		for scanner.Scan() {
-			domain := fmt.Sprintf("http://%s/wso.php", scanner.Text())
+			domain := fmt.Sprintf("%s", scanner.Text())
 			taskChannel <- domain
 		}
 
@@ -94,6 +124,7 @@ func main() {
 	}()
 
 	db := initDataBase()
+
 	wg := new(sync.WaitGroup)
 
 	for i := 0; i < threadsCount; i++ {
@@ -102,4 +133,5 @@ func main() {
 	}
 
 	wg.Wait()
+	db.Close()
 }
